+++ changelog.md
@@ -44,6 +44,7 @@
 - Added info in admin footer for pages provided by the plugin or for which it makes extensions
 - Added support for plugin "WP Extra File Types" to enabled additional possible file types to use as external files
 - Added option to load upload directory via local service
+- Added option to use our plugin name in each HTTP-header User Agent (default enabled)
 
 ### Changed